refactor: rename some error codes and packages
