+++ src/language/validation/other/expressions/templateStrings.ts
@@ -1,7 +1,7 @@
 import {isSdsTemplateStringPart, SdsTemplateString} from "../../../generated/ast.js";
 import {ValidationAcceptor} from "langium";
 
-export const CODE_OTHER_MISSING_TEMPLATE_EXPRESSION = 'other/missing-template-expression';
+export const CODE_TEMPLATE_STRING_MISSING_TEMPLATE_EXPRESSION = 'template-string/missing-template-expression';
 
 export const templateStringMustHaveExpressionBetweenTwoStringParts = (node: SdsTemplateString, accept: ValidationAcceptor): void => {
     for (let i = 0; i < node.expressions.length - 1; i++) {
@@ -11,7 +11,7 @@ export const templateStringMustHaveExpressionBetweenTwoStringParts = (node: SdsT
         if (isSdsTemplateStringPart(first) && isSdsTemplateStringPart(second)) {
             accept('error', 'There must be an expression between two string parts.', {
                 node: second,
-                code: CODE_OTHER_MISSING_TEMPLATE_EXPRESSION,
+                code: CODE_TEMPLATE_STRING_MISSING_TEMPLATE_EXPRESSION,
             });
         }
     }
