+++ src/language/validation/other/statements/assignments.ts
@@ -1,7 +1,7 @@
 import { isSdsPipeline, SdsYield } from '../../../generated/ast.js';
 import { getContainerOfType, ValidationAcceptor } from 'langium';
 
-export const CODE_OTHER_YIELD_FORBIDDEN_IN_PIPELINE = 'other/yield-forbidden-in-pipeline';
+export const CODE_ASSIGMENT_YIELD_FORBIDDEN_IN_PIPELINE = 'assignment/yield-forbidden-in-pipeline';
 
 export const yieldMustNotBeUsedInPipeline = (node: SdsYield, accept: ValidationAcceptor): void => {
     const containingPipeline = getContainerOfType(node, isSdsPipeline);
@@ -9,7 +9,7 @@ export const yieldMustNotBeUsedInPipeline = (node: SdsYield, accept: ValidationA
     if (containingPipeline) {
         accept('error', 'Yield must not be used in a pipeline.', {
             node,
-            code: CODE_OTHER_YIELD_FORBIDDEN_IN_PIPELINE,
+            code: CODE_ASSIGMENT_YIELD_FORBIDDEN_IN_PIPELINE,
         });
     }
 };
