+++ tests/resources/validation/style/unnecessary body in class/main.sdstest
@@ -1,4 +1,4 @@
-package tests.validation.unnecessarySyntax.unnecessaryBodyInClass
+package tests.validation.style.unnecessaryBodyInClass
 
 // $TEST$ info "This body can be removed."
 class MyClass1 »{}«
