+++ tests/resources/validation/style/unnecessary body in enum/main.sdstest
@@ -1,4 +1,4 @@
-package tests.validation.unnecessarySyntax.unnecessaryBodyInEnum
+package tests.validation.style.unnecessaryBodyInEnum
 
 // $TEST$ info "This body can be removed."
 enum MyEnum1 »{}«
