+++ tests/resources/validation/style/unnecessary result list in segment/main.sdstest
@@ -1,4 +1,4 @@
-package tests.validation.unnecessarySyntax.unnecessaryResultListInSegment
+package tests.validation.style.unnecessaryResultListInSegment
 
 // $TEST$ info "This result list can be removed."
 segment mySegment1() »-> ()« {}
