+++ tests/resources/validation/style/unnecessary type parameter list in class/main.sdstest
@@ -1,4 +1,4 @@
-package tests.validation.unnecessarySyntax.unnecessaryTypeParameterListInClass
+package tests.validation.style.unnecessaryTypeParameterListInClass
 
 // $TEST$ info "This type parameter list can be removed."
 class MyClass1»<>«
